Removed unnecessary files. Moved example into example directory
