+++ src/CMakeLists.txt
@@ -1,7 +1,7 @@
-add_library(wrap_me SHARED sampleLib.c)
+add_library(wrap_me SHARED example/sampleLib.c)
 add_library(mein_got SHARED  gotcha_utils.c gotcha.c)
 target_link_libraries(wrap_me mein_got)
-add_executable(symb_look symbolLookup.c)
+add_executable(symb_look example/symbolLookup.c)
 target_link_libraries(symb_look wrap_me mein_got)
 include_directories(${CMAKE_CURRENT_SOURCE_DIR})
 install(TARGETS mein_got DESTINATION lib)
