+++ src/example/symbolLookup.c
